refactor(clawdstrike-policy-event): extract serialize from edr/receipt/mod.rs
